Fixed code per feedback and reformatted to modular programming.
